refactor borshify etc to a single place
